+++ pytest_reportportal/plugin.py
@@ -17,7 +17,7 @@ import logging
 import os.path
 import time
 from logging import Logger
-from typing import Any, Callable, Dict, Generator
+from typing import Any, Callable, Generator
 
 import _pytest.logging
 import dill as pickle
@@ -254,6 +254,7 @@ def pytest_runtest_protocol(item: Item) -> Generator[None, Any, None]:
         filter_client_logs=True,
         endpoint=agent_config.rp_endpoint,
         ignored_record_names=("reportportal_client", "pytest_reportportal"),
+        custom_levels=agent_config.rp_log_custom_levels,
     )
     log_format = agent_config.rp_log_format
     if log_format:
@@ -410,7 +411,7 @@ if PYTEST_BDD:
         scenario: Scenario,
         step: Step,
         step_func: Callable[..., Any],
-        step_func_args: Dict[str, Any],
+        step_func_args: dict[str, Any],
     ) -> Generator[None, Any, None]:
         """Report BDD step finish.
 
@@ -439,7 +440,7 @@ if PYTEST_BDD:
         scenario: Scenario,
         step: Step,
         step_func: Callable[..., Any],
-        step_func_args: Dict[str, Any],
+        step_func_args: dict[str, Any],
         exception,
     ) -> Generator[None, Any, None]:
         """Report BDD step error.
@@ -600,6 +601,12 @@ def pytest_addoption(parser) -> None:
         "rp_log_batch_payload_size",
         help="DEPRECATED: Maximum payload size in bytes of async batch log requests",
     )
+    parser.addini(
+        "rp_log_custom_levels",
+        type="args",
+        help="Custom log levels specified as 'int level:string'. E.G.: '35:ASSERTION'. Overrides existing level if int"
+        " level matches.",
+    )
     parser.addini("rp_ignore_attributes", type="args", help="Ignore specified pytest markers, i.e parametrize")
     parser.addini(
         "rp_is_skipped_an_issue", default=True, type="bool", help="Treat skipped tests as required investigation"
@@ -646,7 +653,6 @@ def pytest_addoption(parser) -> None:
         "directory with certificates of trusted CAs.",
     )
     parser.addini("rp_issue_id_marks", type="bool", default=True, help="Add tag with issue id to the test")
-    parser.addini("retries", default="0", help="Deprecated: use `rp_api_retries` instead")
     parser.addini("rp_api_retries", default="0", help="Amount of retries for performing REST calls to RP server")
     parser.addini(
         "rp_launch_timeout",